use pedal 2 dial 1 offset when writing a program
The first of the two bytes (LE, so the least significant one) seems to have a special meaning in the sysex protocol.
If it has special values above 128 (0x80), its apparently interpreted as a "status byte". Discovered in this scenario:

The dial was set to 749ms, which in LE is 0xED02. Chromium refused to send the WriteUserProgramMessage, stating:

> The system exclusive message contains a status byte at offset 55 (237)

So it appears that the offset is used to prevent this clash of application-level semantics and protocol/sysex
level semantics.